ragljx_py/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# LlamaIndex
llama-index==0.9.48
llama-index-vector-stores-qdrant==0.1.3

# Qdrant
qdrant-client==1.7.0

# MinIO
minio==7.2.0

# OpenAI (for embeddings and LLM)
openai==1.6.1

# Document parsing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.4
markdown==3.5.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

